src/test/java/com/rloup/leetcode/problems/CheckDistanceBetweenSameLettersTest.java
package com.rloup.leetcode.problems;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;

class CheckDistanceBetweenSameLettersTest {

    @Test
    public void checkDistancesTest() {
        CheckDistanceBetweenSameLetters obj = new CheckDistanceBetweenSameLetters();
        boolean b = obj.checkDistances("abaccb", new int[]{1, 3, 0, 5, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0});
        boolean b1 = obj.checkDistances("aa", new int[]{1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0});
        boolean b2 = obj.checkDistances("abbccddeeffgghhiijjkkllmmnnooppqqrrssttuuvvwwxxyyzza",
                                        new int[]{49, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0});

        assertTrue(b);
        assertFalse(b1);
        assertFalse(b2);
    }

}